Move capabilities
